[refactor] move add on entities
