+++ src/config/constants/default_config.ts
@@ -1,5 +1,5 @@
 import { LocalConfig } from "@/domain/service/i_config.service";
-import { AddOnType } from "src/domain/entity/add_on.entity";
+import { AddOnType } from "@/domain/entity/add_on/add_on.entity";
 import { Config } from "src/domain/entity/config.entity";
 import { ObjectType } from "src/domain/entity/dev_object.entity";
