+++ src/domain/entity/add_on/add_on.entity.ts
@@ -1,9 +1,10 @@
-import { AddOnConfig } from "../value_object/add_on_config.vo";
+import { AddOnConfig } from "../../value_object/add_on_config.vo";
 
 export enum AddOnType {
   Analyzer,
   Morpher,
   GitPlatform,
+  Workspace,
 }
 
 export class AddOn {
