+++ src/domain/entity/add_on/analyzer.entity.ts
@@ -1,5 +1,5 @@
 import { AddOn } from "./add_on.entity";
-import { DevObject } from "./dev_object.entity";
+import { DevObject } from "../dev_object.entity";
 import { Morpher } from "./morpher.entity";
 
 export abstract class Analyzer extends AddOn {
