+++ src/domain/entity/add_on/git_platform.entity.ts
@@ -1,7 +1,5 @@
+import { Issue } from "../issue.entity";
 import { AddOn } from "./add_on.entity";
-import { DevObject } from "./dev_object.entity";
-import { Issue } from "./issue.entity";
-import { Morpher } from "./morpher.entity";
 
 export abstract class GitPlatform extends AddOn {
   abstract createIssue(issue: Issue): Promise<Issue>;
