+++ src/domain/entity/add_on/morpher.entity.ts
@@ -1,7 +1,7 @@
-import { AddOnConfig } from "../value_object/add_on_config.vo";
+import { AddOnConfig } from "../../value_object/add_on_config.vo";
 import { AddOn } from "./add_on.entity";
-import { ComponentType } from "./component.entity";
-import { INode } from "./i_node";
+import { ComponentType } from "../component.entity";
+import { INode } from "../i_node";
 
 export interface MorpherConfig extends AddOnConfig {}
 export abstract class Morpher extends AddOn {
