+++ src/domain/service/git_platform.service.ts
@@ -4,8 +4,8 @@ import fs from "fs";
 import { inject, injectable } from "inversify";
 import { GitHub } from "plug_in/git_platform/git_hub";
 import { promisify } from "util";
-import { AddOnType } from "../entity/add_on.entity";
-import { GitPlatform } from "../entity/git_platform.entity";
+import { AddOnType } from "../entity/add_on/add_on.entity";
+import { GitPlatform } from "../entity/add_on/git_platform.entity";
 import { Issue } from "../entity/issue.entity";
 import { IGitPlatformService, IssueWithPR } from "../entity/i_git_platform.service";
 import { AddOnConfig } from "../value_object/add_on_config.vo";
