+++ src/domain/service/morph.service.ts
@@ -1,7 +1,7 @@
 import { injectable } from "inversify";
 import { TypescriptMorpher } from "plug_in/morpher/typescript_morpher";
-import { AddOnType } from "../entity/add_on.entity";
-import { Morpher } from "../entity/morpher.entity";
+import { AddOnType } from "../entity/add_on/add_on.entity";
+import { Morpher } from "../entity/add_on/morpher.entity";
 import { AddOnConfig } from "../value_object/add_on_config.vo";
 import { IMorphService } from "./i_morph.service";
