+++ src/domain/value_object/add_on_config.vo.ts
@@ -1,4 +1,4 @@
-import { AddOnType } from "../entity/add_on.entity";
+import { AddOnType } from "../entity/add_on/add_on.entity";
 import { ObjectType } from "../entity/dev_object.entity";
 
 export interface Property {
